added query me, and it works
I tested the me query and context wasn't working so i traced it and edited a lot of files to try to figure out why it was working. in the process i removed some extra comments that were old code that were no longer needed and some old console logs. im sure theres still more but now we have a reference to me. we need to add roles to the encription so that way we can differ clients from users.