Feature WORKING
added product category types to the database and restricted what it can and can not be. i also set a default of an unnamed category to other so things will stil show up on show all without bogging the other category with uncategorized items. i added conditional rendering to the react side of things to ensure that when you click on the links the approprate products show.